backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
crewai==0.1.0
redis==5.0.1
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
pydantic-settings==2.1.0
weave==0.50.1
openai==1.3.0
exa_py==0.0.40 